medvqa (CLIP Model)/requirements.txt
torch
torchvision
transformers>=4.40.0
datasets
tqdm
numpy
Pillow
gradio
scikit-learn
torchxrayvision
